pyproject.toml
[tool.poetry]
name = "drf-04"
version = "0.1.0"
description = ""
authors = ["Евгений Нестеренко <hectep@gdematerial.ru>"]

[tool.poetry.dependencies]
python = "^3.6"
djangorestframework = "^3.11.0"
django-rest-auth = "^0.9.5"
django-allauth = "^0.41.0"
django-registration = "^3.1"
django-crispy-forms = "^1.8.1"
django-webpack-loader = "^0.6.0"
mixer = "^6.1.3"
pytest-django = "^3.8.0"
psycopg2 = "^2.8.4"
dj-database-url = "^0.5.0"
gunicorn = "^20.0.4"
whitenoise = "^5.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
